Address PR's comments
